@@ -24,9 +24,9 @@ internal class FilesystemRepository(private val file: String) : Tilstandsendring
 
     override fun tilstandsendringer(fordi: List<String>, etter: LocalDateTime?, ignorerTilstand: List<String>, ignorerFordi: List<String>): List<TilstandsendringDto> {
         return testdata
-            .filter { fordi.isEmpty() || it.fordi.toLowerCase() in fordi.map(String::toLowerCase) }
-            .filter { it.fordi.toLowerCase() !in ignorerFordi.map(String::toLowerCase) }
-            .filter { it.tilTilstand.toLowerCase() !in ignorerTilstand.map(String::toLowerCase) }
+            .filter { fordi.isEmpty() || it.fordi.lowercase() in fordi.map(String::lowercase) }
+            .filter { it.fordi.lowercase() !in ignorerFordi.map(String::lowercase) }
+            .filter { it.tilTilstand.lowercase() !in ignorerTilstand.map(String::lowercase) }
             .filter { etter == null || it.sistegang >= etter }
     }
 
